Temp bodge to make MOSH use the system SSH
use system SSH as this seems to just work.

Will need to patch the nix SSH ideally, and then use an overlay to
replace the ssh in nixpkgs so that mosh and others (e.g. repo and git)
will also utilise this patched ssh. The patch needs to account for LDAP
users and also make sure the it reads the correct config and that it
signs in with all the correct keys etc.. (it shouldn't be requiring to
type a password)